♻️ : Move encode/decoder into separate files
